Added util and reorganized includes/code a bit.
Moved the stylesheet into a new specific directory.
Added a util object that currently allows returning the root directory of the source code and using Qt to merge the path strings with the root directory.
The root code directory is now set as a preprocessor macro by CMake.
Changed back Qt "#include" statements to the standard C-style format.