Added initial support for ImageLayout 2.0.
This new ImageLayout will work on the basis of storing primitives (single-class rectangles) and sub-layouts (also rectangles) instead of pixels. This will allow the layouts to be rendered recursively at arbitrary resolutions and simplifies the structure of the layout itself.
Currently, the support isn't implemented, but technically the infrustructure to start making the change is there.