Layouts can now be visualized at root zoom level.
This can be explicity specified by the layout_blend_view.
TODOs:
This system is messy, and the zoom level should probably be re-implemented.
Also, there's a bug with the rendering, it's off when "zoomed in" but rendering at root level.
Also, there general rendering gets very messy if the layout's size is too small.